feat(aws-sdk): add xray-lite-aws-sdk crate
`xray-lite-aws-sdk` is an extension of `xray-lite` for AWS SDK for Rust.
It provides an `Intercept` for AWS SDK clients to facilitate tracing of
AWS service requests. An `Intercept` can be created from a `Context` via
`ContextExt` trait extension, which is automatically implemented for any
`Context` that satisfies the additional bounds.