feat: make Client a trait
`Client` becomes a trait rather than a struct, which defines the
interface of an X-Ray client. The previous implementation is provided by
`DaemonClient` which communicates with the X-Ray daemon.

`Context` has a new type field `Client` that carries the specific type
of the client that has created the context.

`SubsegmentSession` requires a new type parameter `C` that represents
the `Client` type provided by `Context`.

Also applies the above changes to the doctests and the example.